Move winston-related stuff to logging/main
